requirements.txt
# Core Framework
streamlit==1.29.0
python-dotenv==1.0.0

# Blockchain
web3==6.11.3
eth-account==0.10.0

# Database
supabase==2.0.3
psycopg2-binary==2.9.9

# AI Agent - Multiple Providers
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.1
langchain-groq==0.0.1
openai==1.6.1
anthropic==0.8.1
groq==0.4.1

# Data Visualization
pandas==2.1.4
plotly==5.18.0
altair==5.2.0

# Utilities
requests==2.31.0
pydantic==2.5.3

# Document Parsing
docling==1.0.0  # PDF to structured data conversion
